Saml dsop-kode i dsop-pakke
